refactor(api): voting → src/api modules (Effort 1)
Move the voting data-access slice from src/hooks/queries/voting to
src/api/voting, adding a shared types.ts and per-file queryOptions
factories. No behavior change.
