+++ src/pages/EditionView/tabs/ArtistsTab/SetCard/SetVotingButtons.tsx
@@ -1,8 +1,8 @@
 import { Button } from "@/components/ui/button";
 import { VOTE_CONFIG, VOTES_TYPES, type VoteConfig } from "@/lib/voteConfig";
 import { useFestivalSet } from "../FestivalSetContext";
-import { useUserVotes } from "@/hooks/queries/voting/useUserVotes";
-import { useVote } from "@/hooks/queries/voting/useVote";
+import { useUserVotes } from "@/api/voting/useUserVotes";
+import { useVote } from "@/api/voting/useVote";
 import { useAuth } from "@/contexts/AuthContext";
 import { useVoteCount } from "@/hooks/useVoteCount";
