+++ src/pages/ExploreSetPage/ExploreSetPage.tsx
@@ -6,8 +6,8 @@ import { ExplorePageHeader } from "./components/ExplorePageHeader";
 import { CardStackContainer } from "./components/CardStackContainer";
 import { VotingSection } from "./components/VotingSection";
 import { useAuth } from "@/contexts/AuthContext";
-import { useVote } from "@/hooks/queries/voting/useVote";
-import { useUserVotes } from "@/hooks/queries/voting/useUserVotes";
+import { useVote } from "@/api/voting/useVote";
+import { useUserVotes } from "@/api/voting/useUserVotes";
 import { useState } from "react";
 import { useExplorableSets } from "./useExplorableSets";
 import { PageTitle } from "@/components/PageTitle/PageTitle";
