+++ src/pages/SetDetails/SetGroupVoting.tsx
@@ -10,7 +10,7 @@ import {
 import { Badge } from "@/components/ui/badge";
 import { useAuth } from "@/contexts/AuthContext";
 import { useUserGroupsQuery } from "@/hooks/queries/groups/useUserGroups";
-import { useGroupVotesQuery } from "@/hooks/queries/voting/useGroupVotes";
+import { useGroupVotesQuery } from "@/api/voting/useGroupVotes";
 import { Users } from "lucide-react";
 import { VOTE_CONFIG, VOTES_TYPES, getVoteConfig } from "@/lib/voteConfig";
 import { cn } from "@/lib/utils";
